// src/http/generated/models/EachMetadataFilterItems.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * MonoWatch BackBone
 * MonoWatch Security Asset Management
 * OpenAPI spec version: 1.0.0
 */
import type { EachGenericFilter } from './EachGenericFilter';

export interface EachMetadataFilterItems {
  label: string;
  param: string;
  items: EachGenericFilter[];
}